# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TorobRent API
  version: 1.0.0
  description: TorobRent — a smart rental search platform that aggregates, normalizes,
    and ranks property listings from multiple sources.
paths:
...
  /api/v1/messages/{message_id}/:
    get:
      operationId: v1_messages_retrieve
      summary: Open a Message Center item
      parameters:
      - in: path
        name: message_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDetail'
          description: ''
    patch:
      operationId: v1_messages_partial_update
      summary: Change the read state of a Message Center item
      parameters:
      - in: path
        name: message_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMessageReadUpdate'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDetail'
          description: ''
...
components:
  schemas:
...
    ListingInquiryContext:
      type: object
      properties:
        opening_snapshot:
          $ref: '#/components/schemas/ListingInquiryOpeningSnapshot'
        current_availability:
          $ref: '#/components/schemas/ListingInquiryCurrentAvailability'
      required:
      - current_availability
      - opening_snapshot
    ListingInquiryCounterpart:
      type: object
      properties:
        display_name:
          type: string
        role:
          $ref: '#/components/schemas/ListingInquiryCounterpartRoleEnum'
        identity_verified:
          type: boolean
        deleted:
          type: boolean
      required:
      - deleted
      - display_name
      - identity_verified
      - role
    ListingInquiryCounterpartRoleEnum:
      enum:
      - renter
      - submitter
      type: string
      description: |-
        * `renter` - renter
        * `submitter` - submitter
...
    ListingInquiryCurrentAvailability:
      type: object
      properties:
        is_active:
          type: boolean
        state:
          type: string
      required:
      - is_active
      - state
...
    ListingInquiryOpeningSnapshot:
      type: object
      properties:
        property_title:
          type: string
        area_sqm:
          type: integer
        rental_terms:
          $ref: '#/components/schemas/ListingInquiryRentalTermsSnapshot'
        source_display_name:
          type: string
      required:
      - area_sqm
      - property_title
      - rental_terms
      - source_display_name
    ListingInquiryRentalTermsSnapshot:
      type: object
      properties:
        deposit_rial:
          type: integer
        monthly_rent_rial:
          type: integer
        currency:
          type: string
      required:
      - currency
      - deposit_rial
      - monthly_rent_rial
...
    MessageDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        kind:
          allOf:
          - $ref: '#/components/schemas/MessageKindEnum'
          readOnly: true
        title:
          type: string
          readOnly: true
        preview:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        read:
          type: boolean
          readOnly: true
        group:
          allOf:
          - $ref: '#/components/schemas/MessageGroup'
          readOnly: true
        body:
          type: string
          readOnly: true
        target:
          oneOf:
          - $ref: '#/components/schemas/MessageTarget'
          - type: 'null'
          readOnly: true
        public_status:
          type:
          - string
          - 'null'
          readOnly: true
        reply_allowed:
          type: boolean
          readOnly: true
        entries:
          type: array
          items:
            $ref: '#/components/schemas/SupportThreadEntry'
          readOnly: true
        counterpart:
          oneOf:
          - $ref: '#/components/schemas/ListingInquiryCounterpart'
          - type: 'null'
          readOnly: true
        listing_context:
          oneOf:
          - $ref: '#/components/schemas/ListingInquiryContext'
          - type: 'null'
          readOnly: true
        reply_unavailable_reason:
          readOnly: true
          oneOf:
          - $ref: '#/components/schemas/ReplyUnavailableReasonEnum'
          - $ref: '#/components/schemas/NullEnum'
      required:
      - body
      - counterpart
      - created_at
      - entries
      - group
      - id
      - kind
      - listing_context
      - preview
      - public_status
      - read
      - reply_allowed
      - reply_unavailable_reason
      - target
      - title
    MessageGroup:
      type: object
      properties:
        kind:
          $ref: '#/components/schemas/MessageGroupKindEnum'
        id:
          type: string
          format: uuid
        label:
          type: string
      required:
      - id
      - kind
      - label
    MessageGroupKindEnum:
      enum:
      - submission
      - source_proposal
      - support_request
      - listing_inquiry
      type: string
      description: |-
        * `submission` - submission
        * `source_proposal` - source_proposal
        * `support_request` - support_request
        * `listing_inquiry` - listing_inquiry
    MessageKindEnum:
      enum:
      - source_conversation
      - system_notification
      - listing_inquiry
      - support_request
      type: string
      description: |-
        * `source_conversation` - Source Conversation
        * `system_notification` - System Notification
        * `listing_inquiry` - Listing Inquiry
        * `support_request` - Support Request
...
    MessageTarget:
      type: object
      properties:
        label:
          type: string
        href:
          type: string
      required:
      - href
      - label
...
    NullEnum:
      type: 'null'
...
    PatchedMessageReadUpdate:
      type: object
      properties:
        read:
          type: boolean
...
    ReplyUnavailableReasonEnum:
      enum:
      - account_deleted
      - account_blocked
      - listing_inactive
      - responsibility_changed
      type: string
      description: |-
        * `account_deleted` - Account deleted
        * `account_blocked` - Account blocked
        * `listing_inactive` - Listing inactive
        * `responsibility_changed` - Responsibility changed
...
    SupportThreadEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/SupportThreadEntryKindEnum'
        body:
          type: string
        status:
          $ref: '#/components/schemas/SupportThreadEntryStatusEnum'
        created_at:
          type: string
          format: date-time
        edited_at:
          type:
          - string
          - 'null'
          format: date-time
        editable:
          type: boolean
        author_name:
          type: string
        mine:
          type: boolean
      required:
      - created_at
      - id
      - kind
    SupportThreadEntryKindEnum:
      enum:
      - requester_message
      - operator_reply
      - renter_message
      - submitter_message
      - status
      type: string
      description: |-
        * `requester_message` - requester_message
        * `operator_reply` - operator_reply
        * `renter_message` - renter_message
        * `submitter_message` - submitter_message
        * `status` - status
    SupportThreadEntryStatusEnum:
      enum:
      - received
      - in_progress
      - resolved
      type: string
      description: |-
        * `received` - received
        * `in_progress` - in_progress
        * `resolved` - resolved
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid